+++ src/id_translation/dio/__init__.py
@@ -24,6 +24,20 @@ snippet below shows how the :mod:`bundled <.integration>` integrations are regis
 
 The :func:`loader <id_translation.dio.load_integrations>` will skip the integration if calling
 :class:`EntryPoint.load() <importlib.metadata.EntryPoint>` raises an :py:class:`ImportError`.
+
+Selection process
+-----------------
+The :class:`~id_translation.Translator` will call :func:`.resolve_io` once per task. The first implementation whose
+:meth:`DataStructureIO.handles_type`-method returns ``True`` will be used. The order in which implementations are
+considered is determined by the :attr:`~DataStructureIO.priority` attribute.
+
+Bundled implementations have priorities in the `1000 - 1999` range (inclusive); see the table below.
+
+.. csv-table:: Ranking of built-in :class:`DataStructureIO` implementations.
+   :file: /../../tests/dio/expected-order.csv
+   :header-rows: 1
+
+New implementations default to ``priority=10_000`` and are therefore considered first.
 """
 
 from ._data_structure_io import DataStructureIO
@@ -51,8 +65,6 @@ __all__ = [
     "register_io",
     "resolve_io",
 ]
-
-load_integrations()
 
 if __doc__:
     __doc__ = __doc__.format(entrypoint_group=ENTRYPOINT_GROUP)