+++ src/id_translation/dio/_resolve.py
@@ -1,26 +1,17 @@
 import logging
 from typing import Any
 
-from rics.misc import get_public_module, tname
+from rics.misc import tname
 
 from ..types import IdType, NameType, SourceType, TranslatableT
 from ._data_structure_io import DataStructureIO
-from ._dict import DictIO
-from ._pandas import PandasIO
-from ._sequence import SequenceIO
-from ._set import SetIO
-from ._single_value import SingleValueIO
+from .default import DictIO, PandasIO, SequenceIO, SetIO, SingleValueIO
 from .exceptions import UntranslatableTypeError
 
 AnyDataStructureIO = DataStructureIO[Any, Any, Any, Any]
 
-_DEFAULT_IMPLEMENTATIONS: list[type[AnyDataStructureIO]] = [
-    DictIO,
-    SetIO,
-    PandasIO,
-    SequenceIO,
-    SingleValueIO,
-]
+
+_DEFAULT_IMPLEMENTATIONS = PandasIO, DictIO, SetIO, SequenceIO, SingleValueIO
 _RESOLUTION_ORDER: list[type[AnyDataStructureIO]] = [*_DEFAULT_IMPLEMENTATIONS]
 LOGGER = logging.getLogger(__package__)
 
@@ -36,13 +27,11 @@ Background:
 
 def resolve_io(
     arg: TranslatableT,
-    **kwargs: Any,
 ) -> DataStructureIO[TranslatableT, NameType, SourceType, IdType]:
     """Get an IO instance for `arg`.
 
     Args:
         arg: An argument to get IO for.
-        **kwargs: Keyword arguments for the IO class.
 
     Returns:
         A data structure IO instance for `arg`.
@@ -55,7 +44,7 @@ def resolve_io(
     """
     for io_class in _RESOLUTION_ORDER:
         if io_class.handles_type(arg):
-            return _initialize(arg, io_class, kwargs)
+            return _initialize(arg, io_class)
 
     raise UntranslatableTypeError(type(arg))
 
@@ -63,66 +52,75 @@ def resolve_io(
 def _initialize(
     arg: TranslatableT,
     io_class: type[DataStructureIO[TranslatableT, NameType, SourceType, IdType]],
-    kwargs: dict[str, Any],
 ) -> DataStructureIO[TranslatableT, NameType, SourceType, IdType]:
-    try:
-        io = io_class(**kwargs)
-        with_kwargs = True
-    except TypeError:
-        io = io_class()
-        with_kwargs = False
     if LOGGER.isEnabledFor(logging.DEBUG):
-        from rics.misc import format_kwargs
+        LOGGER.debug(
+            f"Using rank-{_RESOLUTION_ORDER.index(io_class)} (priority={io_class.priority}) implementation"
+            f" '{_pretty_io_name(io_class)}' for translatable of type='{tname(arg, include_module=True)}'."
+        )
 
-        pretty_arg = tname(arg, prefix_classname=True)
-        pretty_class = f"{_pretty_io_name(io_class)}({format_kwargs(kwargs)})" if with_kwargs else ""
-        index = _RESOLUTION_ORDER.index(io_class)
-        LOGGER.debug(f"Using rank-{index} implementation {pretty_class} for translatable of type='{pretty_arg}'.")
-    return io
+    return io_class()
 
 
 def get_resolution_order(*, real: bool = False) -> list[type[AnyDataStructureIO]]:
-    """Returns known :class:`.DataStructureIO` implementations.
+    """Returns known :class:`.DataStructureIO` implementations in the correct resolution order.
 
     Args:
         real: If ``True``, return the actual list instead of a copy.
 
     Returns:
-        A list of IO implementations.
+        A list of IO implementations sorted by rank.
     """
-    return _RESOLUTION_ORDER if real else list(_RESOLUTION_ORDER)
+    return _RESOLUTION_ORDER if real else [*_RESOLUTION_ORDER]
 
 
-def register_io(io: type[AnyDataStructureIO]) -> None:
+def register_io(
+    io: type[AnyDataStructureIO],
+) -> None:
     """Register a new IO implementation.
 
-    Classes are polled through :meth:`.DataStructureIO.handles_type` in reverse insertion order (new implementations are
-    polled first). Re-registering an implementation again will move it to the first position in the search order.
+    Classes are polled through :meth:`.DataStructureIO.handles_type` in based on :attr:`DataStructureIO.priority`.
 
     Args:
-        io: A :class:`.DataStructureIO` type.
+        io: A :class:`.DataStructureIO` type
     """
-    if io in _RESOLUTION_ORDER:
-        _RESOLUTION_ORDER.remove(io)
+    if io.priority < 0:
+        LOGGER.warning(f"Refusing to register '{_pretty_io_name(io)}' since priority={io.priority} < 0.")
+        return
 
-    _RESOLUTION_ORDER.insert(0, io)
+    _RESOLUTION_ORDER.append(io)
+
     if LOGGER.isEnabledFor(logging.DEBUG):
-        LOGGER.debug(f"Registered custom IO implementation: '{_pretty_io_name(io)}'.")
+        LOGGER.debug(f"Registered IO implementation: '{_pretty_io_name(io)}'.")
+
+    _finalize_order()
+
+
+def _finalize_order() -> None:
+    new_order = set()
+    for dio in _RESOLUTION_ORDER:
+        if dio.priority < 0:
+            continue
+        new_order.add(dio)
+
+    _RESOLUTION_ORDER.clear()
+    _RESOLUTION_ORDER.extend(new_order)
+    _RESOLUTION_ORDER.sort(key=lambda io: io.priority, reverse=True)
 
 
 def _pretty_io_name(io: type[AnyDataStructureIO]) -> str:
-    return get_public_module(io, resolve_reexport=True) + "." + tname(io, prefix_classname=True)
+    return tname(io, prefix_classname=True, include_module=True)
 
 
 def is_registered(io: type[AnyDataStructureIO]) -> bool:
     """Return IO implementation registration status.
 
-    Instances should register themselves using :func:`register_io` or :meth:`.DataStructureIO.is_registered`.
+    Implementations should register themselves using :meth:`.DataStructureIO.register`.
 
     Args:
         io: A :class:`.DataStructureIO` type.
     """
-    return io in _RESOLUTION_ORDER
+    return io.priority > 0 and io in _RESOLUTION_ORDER
 
 
 def load_integrations() -> None:
@@ -162,12 +160,16 @@ def load_integrations() -> None:
             msg = f"Bad entrypoint={ep!r}: {cls} is not a subtype of {DataStructureIO.__name__}. "
             raise TypeError(msg)
 
-        cls.register()
+        _RESOLUTION_ORDER.append(cls)
         n_ok += 1
+
+    _finalize_order()
 
     millis = round(1000 * (perf_counter() - start))
     LOGGER.debug("Imported and registered %i/%i integrations in %i milliseconds.", n_ok, n_total, millis)
 
+
+load_integrations()
 
 if load_integrations.__doc__:
     load_integrations.__doc__ = load_integrations.__doc__.format(_ENTRYPOINT_GROUP=_ENTRYPOINT_GROUP)