+++ tests/dio/integration/test_dask.py
@@ -9,8 +9,7 @@ from id_translation.dio.integration.dask import DaskIO
 from id_translation.types import IdTypes
 
 assert DaskIO.is_registered(), "entrypoint loader failed"
-DaskIO.register()
-assert DaskIO.get_rank() == 0
+assert DaskIO.get_rank() == 2
 
 
 def to_uuid(i: int) -> UUID: